move vectorize meta to background
