+++ core/Makefile
@@ -1,10 +1,10 @@
 SQLX_OFFLINE:=true
-DATABASE_URL:=postgres://${USER}:${USER}@localhost:28815/postgres
+DATABASE_URL:=postgres://${USER}:${USER}@localhost:28817/postgres
 RUST_LOG:=debug
 
 
 sqlx-cache:
-	cargo sqlx prepare
+	DATABASE_URL=${DATABASE_URL} cargo sqlx prepare
 
 run-worker:
 	cargo run --bin vectorize-worker
